requirements.txt
numpy
streamlit
# streamlit_webrtc
opencv-python-headless==4.5.4.60
opencv-contrib-python-headless==4.5.4.60
click
# av
dlib-binary